refactor.코드 리팩토링 및 구조 변경
