+++ src/components/globeSearch/GlobeSearch.SearchBox.tsx
